UBUNTU: SAUCE: gpio-tegra: Add Tegra234 pin ranges for few PCIE pins
BugLink: https://bugs.launchpad.net/bugs/2001557

Dynamic pinmuxing is needed for PCIE usecase where
pins needs to switch between SFIO and GPIO for RP
and EP mode.

Adding the pins as pin ranges for the pin controller
allows the pin controller to be involved when these
pins are requested as GPIOs and allows the proper
programming to take place.

Signed-off-by: va <va@nvidia.com>
Signed-off-by: Prathamesh Shete <pshete@nvidia.com>
Signed-off-by: Kartik <kkartik@nvidia.com>
Acked-by: Brad Figg <bfigg@nvidia.com>
Acked-by: Jacob Martin <jacob.martin@canonical.com>
Acked-by: Ian May <ian.may@canonical.com>
Signed-off-by: Brad Figg <bfigg@nvidia.com>