ANDROID: KVM: arm64: Do not allow memslot changes after first VM run under pKVM
As the guest stage-2 page-tables will soon be managed entirely by EL2
when pKVM is enabled, guest memory will be pinned and the MMU notifiers
in the host will be unable to reconfigure mappings at EL2 other than
destrroying the guest and reclaiming all of the memory.

Forbid memslot move/delete operations for VMs that have run under pKVM,
returning -EPERM to userspace if such an operation is requested.

Bug: 357781595
Change-Id: I73650c1ac79d8c116a3f31d17ef2a4ef1b30a844
Signed-off-by: Fuad Tabba <tabba@google.com>