ANDROID: virtio_balloon: Do not translate reported pages through DMA API
The free-page reporting and hinting queues do not pass arrays of page
addresses (like the basic inflate queue) but instead pass the free page
ranges as buffers. This does not work well with DMA API: The host wants
to know the GPA, not an IOVA.

For these two virtqueues, disable DMA API and pass through buffers untranslated.

Change-Id: Ic08cf0e2aa092944bac0a6dfc8004c766a7eb3af
Bug: 357781595
Signed-off-by: Keir Fraser <keirf@google.com>