ANDROID: KVM: arm64: Add header_page userspace descriptor for hyp tracing
Added userspace file that describes the content of each hyp tracing
ring-buffer page. This follows the same format as the descriptor found
in the tracefs root to ensure compatibility with userspace trace tools.

We sadly can't use ring_buffer_print_page_header(). This is because we
can have an uninitialized trace_buffer while this file is readable.

Bug: 357781595
Change-Id: Ideebf37cb285b581c4cbe2a991a0df8c7ed61c0f
Signed-off-by: Vincent Donnefort <vdonnefort@google.com>