KVM: arm64: Add guest share/unshare with the FF-A
This allows a guest to share memory with the secure OS[FF-A] client.
After a guest finishes sharing memory with the hyp it can request to
unshare it. This annotates the pages in the guest stage-2 pagetables and
transitions them from the state: OWNED => SHARED_OWNED on the share
operation.

Change-Id: Ie51d80974cfdf943c7cbc7cfa71a345986f866cb
Signed-off-by: Sebastian Ene <sebastianene@google.com>