KVM: arm64: Share the guest FF-A buffers with Secure OS
The guest OS writes the descriptor table to the TX buffers mapped in
the hypervisor. The descriptor table contains a set of constituents
that hold the address and the size of the buffers to be shared with
Trustzone. Copy the guest buffer inside the hypervisor's own buffer and
verify the contents. The hypervisor buffer is the one shared with
Trustzone.

In preparation for supporting Guest memory sharing introduce 4 new
functions which will be implemented in the following patches:
 - ffa_guest_share_ranges / ffa_guest_unshare_ranges
Annotate the Guest stage-2 pagetables that this memory region
is shared/unshared with Trustzone.

 - ffa_guest_repaint_ipa_ranges
For repainting the constituent buffer addresses specified from the Guest
with phisical memory addresses(Note: Trustzone has no information about
our Guest address-space & mappings).

- is_ffa_id_valid
Verifies if the VM is allowed to do FF-A memory operations

Signed-off-by: Sebastian Ene <sebastianene@google.com>