ANDROID: KVM: arm64: Donate memory for the guest SVE state for protected VMs
The host already allocates separate memory for the guest's SVE
state and shares it with hyp. For protected VMs, SVE is
completely handled in hyp and the host has no reason to access
it. Donate it instead.

Note that this donation does not allocate more memory than
earlier, it just donates it instead of sharing it.

Bug: 357781595
Change-Id: Iedc71607cf67c1b4c0a895548e3d7b158ca732b8
Signed-off-by: Fuad Tabba <tabba@google.com>