KVM: arm64: Introducing pkvm_mem_abort_range()
Previously, guest stage-2 mapping was only triggered on a guest fault.
If the hypervisor needs one (e.g to share memory between guest and
host) its only solution is to fake a fault and to return to host. Of
course being done on a per-page basis, it is quite inefficient.

Instead provide a function to map an entire range for a guest stage-2
which will later be used to satisfy hypervisor mapping requests.

Even if a range is passed, this function doesn't allow the creation of
block mappings.

Signed-off-by: Vincent Donnefort <vdonnefort@google.com>