UBUNTU: SAUCE: hwmon: support setting limits in dts for lm90
BugLink: https://bugs.launchpad.net/bugs/2028400

The following sysfs nodes are provided by lm90:
- max limit for local sensor
- max limit for remote sensor
- max limit for second remote sensor
- critical limit for local sensor
- critical limit for remote sensor
- critical limit for second remote sensor
- temperature offset for remote sensor

The above limits can be set from the userspace. However, there may be a
case where at some point before the userspace sets the thresholds, the
board/chip temperature crosses a threshold and then returns back to the
normal range. It will be safer if the thresholds can be set at an
earlier stage so that corresponding actions can be taken. Therefore,
this change adds support for taking the thresholds in the device tree if
any so that they can be set during driver probing.

Signed-off-by: Yi-Wei Wang <yiweiw@nvidia.com>
Reviewed-by: Bitan Biswas <bbiswas@nvidia.com>
Signed-off-by: Abhilash G <abhilashg@nvidia.com>
Acked-by: Jacob Martin <jacob.martin@canonical.com>
Acked-by: Ian May <ian.may@canonical.com>
Signed-off-by: Jacob Martin <jacob.martin@canonical.com>