NVIDIA: SAUCE: hte: modify gpio controller match function
BugLink: https://bugs.launchpad.net/bugs/2056486

The linux jammy kernel does not have fwnode member in the
struct gpio_chip. This gives an error when using the upstream
patch for the tegra234 HTE provider. The CL addresses that
concern and implements match function to match the of_node
instead of the fwnnode

Signed-off-by: Dipen Patel <dipenp@nvidia.com>
Reviewed-by: Gautham Srinivasan <gauthams@nvidia.com>
Reviewed-by: Jon Hunter <jonathanh@nvidia.com>
Signed-off-by: Brad Griffis <bgriffis@nvidia.com>
Acked-by: Jacob Martin <jacob.martin@canonical.com>
Acked-by: Noah Wager <noah.wager@canonical.com>
Signed-off-by: Noah Wager <noah.wager@canonical.com>