NVIDIA: SAUCE: phy: tegra194-p2u: Fix device-tree backward compatibility
BugLink: https://bugs.launchpad.net/bugs/2056594

The change that added PHY lane margining support to the Tegra194 P2U
PHY driver breaks device-tree backward compatibility. In other words,
if you attempt to boot a kernel with the lane margining change and a
device-tree that does not have the 'intr' interrupt node, required by
this change, then the PHY driver fails to probe and the following errors
are seen ...

 tegra194-p2u 3e00000.phy: IRQ intr not found
 tegra194-p2u 3e00000.phy: failed to request "intr" irq
 tegra194-p2u: probe of 3e00000.phy failed with error -22

Any updates to the device-tree bindings must be backward compatible so
that you can boot with an older device-tree. Hence, whenever adding more
features to existing device-tree bindings, we need to ensure that older
bindings still work. Fix the PHY driver so that if the 'intr' interrupt
is not present, then we do not fail to probe, but just disable lane
margining, if supported.

Fixes: "UBUNTU: SAUCE: phy: tegra: p2u: Add lane margin support"
Signed-off-by: Jon Hunter <jonathanh@nvidia.com>
Signed-off-by: Brad Griffis <bgriffis@nvidia.com>
Acked-by: Jacob Martin <jacob.martin@canonical.com>
Acked-by: Noah Wager <noah.wager@canonical.com>
Signed-off-by: Noah Wager <noah.wager@canonical.com>