UBUNTU: SAUCE: memory: tegra: add check if MRQ_EMC_DVFS_LATENCY is supported
BugLink: https://bugs.launchpad.net/bugs/2028400

Add check to ensure that "MRQ_EMC_DVFS_LATENCY" is supported by the
BPMP-FW before making the MRQ request. Currently, if the BPMP-FW
doesn't support this MRQ, then the "tegra186_emc_probe" fails as the
value returned in "response.num_pairs" could be arbitrarily high
resulting in warning from kmalloc.
Due to this the Memory Interconnect initialization also doesn't
happen. Memory Interconnect is not dependent on this MRQ and can
initialize even when this MRQ is not supported. The check ensures
that the MRQ is called only when it is supported by the BPMP-FW
and Interconnect initializes independent of this MRQ.

Also, added the tegra_bpmp_transfer() functions return value check
from the below given upstream patch.
  patch: https://patchwork.kernel.org/project/linux-pci/patch/20210915085517.1669675-3-mperttunen@nvidia.com/

  WARNING: CPU: 1 PID: 8 at __alloc_pages+0x254/0x2c0
  ....
  Call trace:
   __alloc_pages+0x254/0x2c0
   kmalloc_large_node+0x58/0xd0
   __kmalloc_node_track_caller+0x440/0x530
   devm_kmalloc+0x54/0x120
   tegra186_emc_probe+0xe4/0x410

Signed-off-by: Sumit Gupta <sumitg@nvidia.com>
Reviewed-by: Timo Alho <talho@nvidia.com>
Reviewed-by: Bibek Basu <bbasu@nvidia.com>
Signed-off-by: Abhilash G <abhilashg@nvidia.com>
Acked-by: Jacob Martin <jacob.martin@canonical.com>
Acked-by: Ian May <ian.may@canonical.com>
Signed-off-by: Jacob Martin <jacob.martin@canonical.com>