KVM: arm64: Implement guest share/unshare memory region
This annotates the guest stage-2 pagetables marking the pages as being
shared with Trustzone. The unshare removes the annotation and clears out
the sharing attributes.

When the hypervisor returns EFAULT because the page to be shared is not
mapped into guest stage-2, we inject a fault into the host using
the __pkvm_memshare_page_req mechanism and we unshare all the previously
shared pages on the error handling path. The injected fault triggers a
memory donation from the host to the guest which will create the memory
mapping.

Modify the smc handling code for pVMs so that we return to the host
without incrementing the program counter which will allow us to replay
the last instruction after the page is donated to the guest.

Signed-off-by: Sebastian Ene <sebastianene@google.com>