KVM: arm64: Only flip swbits for pKVM guest share/unshare oprations
When a guest shares (or unshares) a page, we need a mapping in place.
This makes sure no subsequent memory fault would override the state of
the mapping. As a consequence, when the page is actually shared, checks
have already enforced the presence of a mapping. At this point, we can
simply mark the sharing status in the software bits instead of remapping
entirely the range.

Signed-off-by: Vincent Donnefort <vdonnefort@google.com>