KVM: arm64: mem_protect to use size instead of nr_pages
A number of pages is not very convinient. First, the actual size might
be different for different entities: Guests could simply use a different
page size. Second, the pgtable code already relies on size and not on a
number of pages, so mem_protect ends-up converting nr_pages to size (and
reciprocally) everywhere. A size argument is then the most natural way
for mem_protect to store a range.

Signed-off-by: Vincent Donnefort <vdonnefort@google.com>