arm64: ptdump: Register a debugfs entry for the host stage-2 page-tables
Initialize the structures used to keep the state of the stage-2 ptdump
walker. To satisfy the ptdump API for parsing no VMA regions, initialize
a memory structure. Since we are going to parse a snapshot of the host
stage-2 paget-tables we don't rely on the locking from the memory
structure.

Signed-off-by: Sebastian Ene <sebastianene@google.com>