ANDROID: KVM: arm64: iommu: Allocate memory for idmapped domains
We need to support devices that don't use the DMA API and expect the
IOMMU in bypass, or that has another stage for IOVA => IPA and use
the second stage just for protection.

For these devices, we keep a page table that track the CPU mappings,
and right now we don't return from contexts requesting memory for IOMMU
donations, so we need a pre-populated pool for the idmapped page table.

Bug: 357781595
Bug: 384432312
Change-Id: Ica9583d8ecb5b2406599ed2cf6f0f758c8ce0cc9
Signed-off-by: Mostafa Saleh <smostafa@google.com>