ANDROID: KVM: arm64: Huge page support for pkvm_pinned_page
A single pin is expected to be held for an entire huge page. However, a
guest can decide (even if that is completely counter-productive) to
relinquish a small portion of that huge page. A new counter (pins) is
added to only release a huge page once its entire range has been
yielded.

Bug: 357781595
Bug: 278011447
Change-Id: I43d49c286a550fcb8ee861d3d9df9f85f4d2bd76
Signed-off-by: Vincent Donnefort <vdonnefort@google.com>