ANDROID: KVM: arm64: Enforce type check for pkvm_register_el2_call()
dyn_hcall_t has changed in between Android releases. With the current
macro, no type check would happen and the incompatibility would even
stay unnoticed without kCFI.

Originally, this declaration was a macro to help with the old
implementation of kCFI. This is not needed anymore, hence use inline
instead.

Bug: 357781595
Bug: 244543039
Bug: 244373730
Change-Id: I0fba347946ba64ab5f9c39e65d17c29839ff83d1
Signed-off-by: Vincent Donnefort <vdonnefort@google.com>