ANDROID: KVM: arm64: Reduce host/hyp vcpu state copying
When running with pKVM enabled, protected guests run with a fixed
CPU configuration and therefore features such as hardware debug
and SVE are unavailable and their state does not need to be
copied from the host structures on each flush operation.

Although non-protected guests do require the host and hyp
structures to be kept in-sync with each other, we can defer
writing back to the host to an explicit sync hypercall, rather
than doing it after every vCPU run.

TODO: check debug state syncing

Bug: 357781595
Change-Id: Ia80ae7bf8e374a50fda4ed5637abdfb82bcf3715
Signed-off-by: Fuad Tabba <tabba@google.com>