ANDROID: KVM: arm64: Check supported features for protected VMs via the value of their feature registers
Do not use feature masks when checking if a protected VM's
supported features are correct. The feature masks  will be
removed in future patches and which might be missing some
information, e.g., if the vcpu itself doesn't support the
feature.

Instead, read the value of the vcpu's feature register directly
to ensure that feature support is correct (or as expected).

Bug: 357781595
Change-Id: Ic63754c9f681b910624767573b1e855bfea7211d
Signed-off-by: Fuad Tabba <tabba@google.com>