ANDROID: KVM: arm64: Introduce and use RESET_ZERO and RESET_UNKNOWN for pvm system registers
To help debug the guest and to detect potential problems with the
guest state, introduce and use RESET_UNKNOWN, which sets the
value of system registers with an architecturally "unknown" value
(according to the Arm ARM), to a value that is easy to recognize.

While we're at it, introduce RESET_ZERO, for consistency with
sysreg code in kvm in general, and to tidy up.

Bug: 357781595
Change-Id: I74f6064efca556d3da8647c184218f89b3ad31ce
Signed-off-by: Fuad Tabba <tabba@google.com>