ANDROID: KVM: arm64: Pass pkvm_el2_module struct to hypervisor
When a pKVM module is initialized, only a pointer to the init function
is give to the hypervisor. With the growing set of features the pKVM
modules are supporting, more and more initialization is necessary in the
hypervisor. Give access to the pkvm_el2_module containing the
addresses of ELF sections to facilitate that job.

Bug: 357781595
Bug: 244543039
Bug: 244373730
Change-Id: Id1595719acfddeb0ae04fe6387f47ee15cc2e64f
Signed-off-by: Vincent Donnefort <vdonnefort@google.com>