ANDROID: KVM: arm64: Mark vcpu as invalid on unhandleable trap
Some trap types cannot be handled for protected VMs, and are not
expected ever to occur. If they do, they are punted to Host with
no associated context, which ends up trying to use a stale ESR.

It is better to fail-stop the affected VCPU in this case, and
emit a trace log entry.

Bug: 357781595
Change-Id: Ida3dd68f67afe0bd7451a4d4b8b038fd96934b1c
Signed-off-by: Keir Fraser <keirf@google.com>