BACKPORT: FROMLIST: KVM: arm64: iommu: Add map/unmap() operations
Handle map(), unmap() and iova_to_phys() hypercalls.

In addition to map/unmap, the hypervisor has to ensure that all
mapped pages are tracked, so be before each map()
__pkvm_host_use_dma() would be called to ensure that.

Similarly, on unmap() we need to decrement the refcount using
__pkvm_host_unuse_dma().

However, doing this in standard way as mentioned in the comments is
challenging, so we leave that to the driver.

Also, the hypervisor only guarantees that there are no races
between alloc/free domain operations using the domain refcount to
avoid using extra locks.

Link: https://lore.kernel.org/all/20241212180423.1578358-19-smostafa@google.com/

Bug: 357781595
Bug: 384432312

Change-Id: I69e3eb5d89692c31e7014a08414f80aba686844b
Signed-off-by: Mostafa Saleh <smostafa@google.com>
Signed-off-by: Jean-Philippe Brucker <jean-philippe@linaro.org>