BACKPORT: FROMLIST: iommu/arm-smmu-v3-kvm: Add host driver for pKVM
Under protected KVM (pKVM), the host does not have access to guest or
hypervisor memory. This means that devices owned by the host must be
isolated by the SMMU, and the hypervisor is in charge of the SMMU.

Introduce the host component that replaces the normal SMMUv3 driver when
pKVM is enabled, and sends configuration and requests to the actual
driver running in the hypervisor (EL2).

Rather than rely on regular driver probe, pKVM directly calls
kvm_arm_smmu_v3_init(), which synchronously finds all SMMUs and hands
them to the hypervisor. If the regular driver is enabled, it will not
find any free SMMU to drive once it gets probed.

Link: https://lore.kernel.org/all/20241212180423.1578358-43-smostafa@google.com/

Bug: 357781595
Bug: 384432312

Change-Id: I86ba6cc3e4db318e675ef45f9099c10789728b42
Signed-off-by: Jean-Philippe Brucker <jean-philippe@linaro.org>
Signed-off-by: Mostafa Saleh <smostafa@google.com>