ANDROID: KVM: arm64: Allow the pVM guest to boot with different granule
Remove the limitation of having the same page size when we probe
the hypervisor services from a protected guest VM and allow 4K
protected guest VMs to boot on 16Kb kernels.

Test: Build Microdroid Kernel with this patch.
tools/bazel run --config=fast --lto=thin
//common:kernel_aarch64_microdroid_dist -- --destdir=out/dist
Replace the kernel with the one already built in the aosp repo:
$(aosp)/package/modules/Virtualization/guest/kernel/android15-6.6/arm64/kernel-6.6
Rebuild virtualization service and reinstall the apex
Reboot the device and make sure that protected VMs start on a 16Kb host.
Bug: 381400679
Bug: 357781595
Change-Id: I8ac96833c3fa031a40f02ca18d786febce8aa3b8
Signed-off-by: Sebastian Ene <sebastianene@google.com>