ANDROID: KVM: arm64: iommu: Avoid use of uninitialized variable in smmu_map_pages
When the while condition is first evaluated, the variable named ret may
be used uninitialized. Additionally, it is not necessary to make it part
of the while condition, because the only place where this variable is
changed is an assignment after the call to map_pages, and there is
already an if statement that checks its value.

Bug: 394152442
Fixes: a737b7d0e721 ("ANDROID: KVM: arm64: iommu: Reduce the logic in generic code")
Change-Id: I3a214f5ac424bc13b2443097cd6a2a05a6397393
Signed-off-by: Daniel Mentz <danielmentz@google.com>
Signed-off-by: Mostafa Saleh <smostafa@google.com>