KVM: arm64: Introduce hypercalls to reclaim guest memory
Memory pages donated to a protected guest VM will need to be reclaimed
by the host upon teardown. Introduce a new 'is_dying' state for a
protected VM in which the host is allowed to reclaim guest pages after
having been poisoned by the hypervisor. The 'is_dying' state is a point
of no return past which no vCPU of the pVM is allowed to run any more.

The hypercall to reclaim pVM pages allows to do so one page at a time
on purpose. The EL2 code being non-preemptible, it is preferable to
return to the EL1 frequently during the teardown of a guest to allow
scheduling of other tasks.

Change-Id: I548263c98b6935d83ada3946f9683b13e837cc39
Signed-off-by: Quentin Perret <qperret@google.com>