ANDROID: KVM: iommu: Add flags to attach_dev
There is no way to pass device specific information at attach time
(to handle errata or special IOMMU configuration).
Add a new argument to attach hypercalls “flags” which can be passed
from the EL1 to EL2 driver to describe any driver specific information.
It ‘s mandatory for devices supporting pvIOMMU, to reserve the value “0”
as this is not exposed to guests and “0” would be passed.

Bug: 357781595
Change-Id: I81a0b82591a0fe6d952f46025fbd95a82020c14d
Signed-off-by: Mostafa Saleh <smostafa@google.com>