ANDROID: KVM: arm64: Use unmap for pKVM guests memory relinquish
Previously, annotate was used when a guest relinquishes memory. The
intension was to allow splitting an existing block. However we are
moving away from this mechanism (which came with a list of serious
issues). Relinquish will now on only happen on level-3 mappings and
therefore we can restore the unmap.

Bug: 419518210
Bug: 419548963
Bug: 357781595
Bug: 278011447
Change-Id: I4e7d2af3683cf2b189b1580f39d543d83cac644f
Signed-off-by: Vincent Donnefort <vdonnefort@google.com>