ANDROID: KVM: arm64: iommu: Add driver init
When the drivers were built-in all the kernel initialization where
done from an init call, then the hypervisor init was done at
__pkvm_init.

With modules, the driver might want to have some initialization
code after all modules are loaded and before de-privilege.

Also, we would need a single point for both configs to init the
hypervisor part before de-privilge and and after KVM init.

Bug: 461757236
Signed-off-by: Mostafa Saleh <smostafa@google.com>