ANDROID: KVM: arm64: iommu: Split init
Instead of relying on init call to do driver init, now it's called
at a certain point after early modules, that is needed for IOMMU
modules.
At the moment the only driver (KVM SMMUv3) is only using the correct
API, but doesn't rely on that as it is not modular yet.

Signed-off-by: Mostafa Saleh <smostafa@google.com>